Correct some errors
